+++ mern_vb_backend/models/Savings.js
@@ -1,3 +1,5 @@
+const mongoose = require('mongoose');
+
 const savingSchema = new mongoose.Schema({
     userId: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true },
     month: { type: Number, required: true },
@@ -5,6 +7,6 @@ const savingSchema = new mongoose.Schema({
     date: { type: Date, default: Date.now },
     fine: { type: Number, default: 0 },
     interestEarned: { type: Number, default: 0 }
-  });
-  
-  module.exports = mongoose.model('Saving', savingSchema);
+});
+
+module.exports = mongoose.model('Saving', savingSchema);